close #10

+++ APS - Analyze Profit of Spaceships for X4/APS/Methods/Profit/ChartShipVM.cs
@@ -3,7 +3,7 @@
     public class ChartShipVM
     {
         public string TimeRounded { get; internal set; }
-        public double EstimatedProfit { get; set; }
+        public decimal EstimatedProfit { get; set; }
         public string Ship { get; internal set; }
         public int Quantity { get; internal set; }
         public int Money { get; set; }
